pyproject.toml
[tool.poetry]
name = "online-learning-platform"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
python-dotenv = "^1.2.1"
django-filter = "^25.2"


[tool.poetry.group.dev.dependencies]
django = "^6.0.1"
psycopg2-binary = "^2.9.11"
pillow = "^12.1.0"
djangorestframework = "^3.16.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
